deploy.shにコミットID付ける

+++ deploy.sh
@@ -1,5 +1,14 @@
 #!/bin/sh
 
+COMMENT=$1
+
+if [ -z ${COMMENT} ]; then # commentのサイズが0であれば真
+  echo "Usage: ./deploy.sh [comment statement(required)]"
+  exit
+fi
+
+COMMIT_HASH=`git log -1 --format=%h`
+
 # pagesフォルダの ., .., .git 以外を消去
 ls -la pages | grep -v -E '.|.git$' | xargs rm -rf
 
@@ -9,5 +18,6 @@ hugo -d pages
 # deploy
 cd pages
 git add .
-git commit -m 'updated by deploy.sh'
+git commit -m "COMMENT: ($COMMENT)" -m "HASH: ($COMMIT_HASH)"
 git push origin gh-pages
+
